第一排的同学依旧很困，第二排的同学没有网，第三排和第四排的同学平安无事

+++ itrip-commons/src/main/java/com/artiecode/itrip/pojo/vo/HotelRoomVO.java
@@ -12,7 +12,7 @@ public class HotelRoomVO implements Serializable{
     private Long id;
     private Long hotelId;
     private String roomTitle;
-    private BigDecimal roomPrice;
+    private Double roomPrice;
     private Long roomBedTypeId;
     private Integer isHavingBreakfast;
     private Integer payType;
@@ -45,11 +45,11 @@ public class HotelRoomVO implements Serializable{
         this.roomTitle = roomTitle;
     }
 
-    public BigDecimal getRoomPrice() {
+    public Double getRoomPrice() {
         return roomPrice;
     }
 
-    public void setRoomPrice(BigDecimal roomPrice) {
+    public void setRoomPrice(Double roomPrice) {
         this.roomPrice = roomPrice;
     }
